Rework a lot of stuff
* Made Level layout be a background texture rather than being game objects
* Rewrote level generator (Old one is now level_.h)
* Pellets are still GameObjects
* Regular and Mega Pellets are now differentiated
* Use less dirty pointers
