+++ Test Bench/Header Files/Components/PelletsComponent.h
@@ -1,9 +1,4 @@
 #pragma once
 
-class Pellet {
-	using Position = bloom::components::Position;
-	public:
-		Pellet(Position a) : PelletPosition(a) {}
-
-		Position PelletPosition;
-};
+struct Pellet {};
+struct MegaPellet {};
