+++ Test Bench/data/Assets/Level/0.txt
@@ -1,3 +1,5 @@
+28 31
+
 1  2  2  2  2  2  2  2  2  2  2  2  2  3  4  2  2  2  2  2  2  2  2  2  2  2  2  5
 6  37 37 37 37 37 37 37 37 37 37 37 37 8  9  37 37 37 37 37 37 37 37 37 37 37 37 6
 6  37 16 14 14 15 37 16 14 14 14 15 37 8  9  37 16 14 14 14 15 37 16 14 14 15 37 6
